update: beta 4 (#474)
* update: beta 4

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>

* 🔧 chore(package.json): move tslib to dependencies
The tslib package is now moved from devDependencies to dependencies. This is because tslib is a runtime dependency and is required for the application to run.

* 🐛 fix(pages_test.yml): use secrets.GITHUB_TOKEN in npm set command
The npm set command now uses the secrets.GITHUB_TOKEN environment variable instead of a hardcoded value. This ensures that the correct token is used for authentication when logging in to Github Packages.

* 🐛 fix(testing.js): change npm run deploy to npm run build
The `npm run deploy` command was changed to `npm run build` to correctly reflect the command that needs to be run.

---------

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>