🔧 chore(action.yml): update main script path to dist/index.js (#477)
🔧 chore(tsconfig.json): update target to ES2021 and add lib option
The main script path in the action.yml file has been updated to point to the compiled code in the dist folder. This ensures that the action can be run without requiring compilation. The target in the tsconfig.json file has been updated to ES2021 to take advantage of the latest features of the ECMAScript standard. The lib option has also been added to specify the library files that should be included in the compilation process.