chore: update to Beta 6 (#481)
* 🔊 chore(logging): add debug logs to several functions
This commit adds debug logs to several functions to help with debugging and troubleshooting. The logs will help to identify the flow of execution and the values of variables at different points in the code.

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>

* 🔥 refactor(config): remove "onColumn" from "enforceConventions" object
The "onColumn" property was removed from the "enforceConventions" object in three different files. This property was not being used and was causing confusion.

* 🔖 chore(package.json): bump version to 1.0.0-beta.6
This commit updates the version of the package to 1.0.0-beta.6. This is a chore commit as it does not introduce any new features or fix any bugs, but rather updates the version number to reflect changes made to the codebase.

* 📝 chore(pr.json): update pull request information

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>

---------

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>